tests/eth2/beacon: move test to beacon-chain-specific file.
`test_validation.py` conflicts with another file by same name, so
move test to `test_beacon_validation.py` that was made just for that,
in PR https://github.com/ethereum/py-evm/pull/1660
