+++ tests/eth2/beacon/test_beacon_validation.py
@@ -6,6 +6,7 @@ from eth_utils import (
 
 from eth2.beacon.validation import (
     validate_slot,
+    validate_slot_for_state_slot,
 )
 
 
@@ -30,3 +31,41 @@ def test_validate_slot(slot, is_valid):
     else:
         with pytest.raises(ValidationError):
             validate_slot(slot)
+
+
+@pytest.mark.parametrize(
+    (
+        'state_slot, slot, epoch_length, success'
+    ),
+    [
+        (
+            0, 0, 64, True,
+        ),
+        (
+            64 * 2, 64, 64, True,
+        ),
+        (
+            64 * 2, 64 - 1, 64, False,  # slot is too small
+        ),
+        (
+            64 * 2, 64 * 3 - 1, 64, True,
+        ),
+        (
+            64 * 2, 64 * 3, 64, False,  # slot is too large
+        ),
+    ]
+)
+def test_validate_slot_for_state_slot(state_slot, slot, epoch_length, success):
+    if success:
+        validate_slot_for_state_slot(
+            state_slot=state_slot,
+            slot=slot,
+            epoch_length=epoch_length
+        )
+    else:
+        with pytest.raises(ValidationError):
+            validate_slot_for_state_slot(
+                state_slot=state_slot,
+                slot=slot,
+                epoch_length=epoch_length
+            )
